requirements.txt
matplotlib==3.5.3
scipy==1.7.3
sklearn
pyyaml
wandb==0.15.5
jupyterlab==1.2.6
numpy==1.21.6
torch==1.13.1
pandas==1.3.5
pytest==7.4.3
